inference-cloud/requirements.txt
aws-cdk-lib==2.121.1
constructs>=10.0.0,<11.0.0
boto3
botocore
